Grish

+++ Meeting-Management-System/src/main/java/Meeting_Management/System/Entity/Gender.java
@@ -10,7 +10,8 @@ import java.time.ZonedDateTime;
 public class Gender {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+
+    private Integer id;
 
     @Column(name = "gender_alias", length = 10, nullable = false, unique = true)
     private String genderAlias;
